ux: prevent overscroll in history view
Overscroll can occur, if you scroll really fast (for example using a
Logitech MX Master mouse with MagSpeed Electromagnetic scrolling) and
then suddenly stop scrolling (by stopping the scroll wheel with your
finger).  If the scrolling itself doesn't stop immediately, you
experience overscroll.  It could also be described as lagging.

The reason is, that the scroll events are buffered and pile up on fast
scrolling.  The solution is to ignore events that would lead to
overscroll.