first variant of commit message wrapping
This first variant supports the following features:

  - converting of tabs to spaces 
  - normalizing line ends
  - using indentation of first line for the complete paragraph
  - wrapping lines after 72 characters
  - typing Enter/Return inserts two newlines (new paragraph)

This is done on-the-fly while typing or editing the commit message.  At
least one caveat remains: you couldn't generate something like this
commit message where the indented lines are not complete paragraphs.  A
differentiation between soft and hard breaks is missing.  At the moment
all breaks are interpreted as soft breaks.