prevent unsolved branches in commit graph
For each merge commit where a parent was not part of the focussed commit
graph, an unsolved open branch from the merge commit would be
represented as an open (non-ended) curve.  This could only be prevented
by using the "First-parent only" option, which is sometimes not wanted.

Now these branches are no longer shown.  The merge itself remains marked
as a merge commit (the plus sign in the dot).